refactor: modified db schema of field ID (#34)

+++ pkg/pagination/page_token.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 )
 
-func DecodePageToken(pageToken string) (int, error) {
+func DecodePageToken(pageToken string) (int64, error) {
 	bytes, err := base64.StdEncoding.DecodeString(pageToken)
 	if err != nil {
 		return 0, errors.New("page token is invalid")
@@ -16,10 +16,10 @@ func DecodePageToken(pageToken string) (int, error) {
 	if err != nil {
 		return 0, errors.New("page token is invalid")
 	}
-	return int(token), nil
+	return token, nil
 }
 
-func EncodePageToken(n int) (string, error) {
+func EncodePageToken(n int64) (string, error) {
 	token := base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%v", n)))
 	return token, nil
 }
